+++ CHANGELOG.md
@@ -7,14 +7,18 @@
 
 ### New features
 
+### Bugfixes
+</details>
+
+## 0.9.1
+
+### New features
+
 - Added `\Wizaplace\SDK\Order\DeclinationOption`
 - Added `\Wizaplace\SDK\Order\OrderItem::getDeclinationOptions`
 - `\Wizaplace\SDK\User\UpdateUserCommand::setBirthday` now takes a `\DateTimeInterface` instead of a `\DateTime`
 - Added `\Wizaplace\SDK\Order\OrderService::sendAfterSalesServiceRequest`
 - Added `\Wizaplace\SDK\Catalog\ProductAttribute::getId`
-
-### Bugfixes
-</details>
 
 ## 0.9.0